Add new packet type: vector-3f
This is a tridimensional (x, y, z) vector of doubles, with attached
min/max values, applied equally to all axes/dimensions.

All node converters to/from other types were added to, along with
proper tests.

The first users of this new packet type will be the accelerometer and
gyroscope nodes.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>